Handle ResourceBusy
On Apple Silicon (M2 Macbook Air), using `DrmOutputManager` causes
`ResourceBusy` error on the first call of
`DrmOutputManager::queue_frame()`. This will be handled as
`TemporaryErorr` and it works if continue to render. To continue the
rendering loop, this patch reschedules rendering on `ResourceBusy`.

See https://github.com/Smithay/smithay/issues/1661 and
https://github.com/Smithay/smithay/pull/1662.